receiveAndSendRequestToSlave package renamed, tests fixed
